♻️ refactor(services/image-manager): Migrate to `.ts` using `SvelteMap`
